third_party/WebKit/LayoutTests/inspector/profiler/cpu-profiler-stopped-removed-race.html

<html>
<head>
  <script src="../../http/tests/inspector/inspector-test.js"></script>
  <script src="profiler-test.js"></script>
<script>

function test()
{
    InspectorTest.runProfilerTestSuite([
        function testProfiling(next)
        {
            var cpuProfiler = WebInspector.targetManager.mainTarget().cpuProfilerModel;
            var targetManager = WebInspector.targetManager;
            targetManager.addEventListener(WebInspector.TargetManager.Events.SuspendStateChanged, onSuspendStateChanged);
            var profilesPanel = WebInspector.panels.profiles;
            InspectorTest.addSniffer(cpuProfiler, "stopRecording", stopRecording);
            InspectorTest.addSniffer(profilesPanel, "_addProfileHeader", onAddProfileHeader);
            profilesPanel.toggleRecordButton();  // Start profiling.

            function onAddProfileHeader()
            {
                profilesPanel.toggleRecordButton();  // Stop profiling.
                profilesPanel._reset();  // ... and remove the profile before it actually stopped.
            }

            function onSuspendStateChanged()
            {
                if (WebInspector.targetManager.allTargetsSuspended()) {
                    InspectorTest.addResult("Suspending targets");
                    return;
                }
                InspectorTest.addResult("Resuming targets");
                InspectorTest.completeProfilerTest();
            }

            function stopRecording(resultPromise)
            {
                InspectorTest.addResult("Stop recording command issued.");
                resultPromise.then(didStopRecording);
            }

            function didStopRecording(profile)
            {
                InspectorTest.addResult("Recording stopped. profile (should be null): " + profile);
            }
        }
    ]);
}

</script>
</head>
<body onload="runTest()">
<p>
Tests that profile removal right after stop profiling issued works.

<a href="http://crbug.com/476430">Bug 476430.</a>
</p>
</body>
</html>
